+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Note: You may need to set "DVD order" in your settings again as the internal set
  - Fix ADE actor scraping (#1011)
  - Use user's preferred UI language and not the system language (#1002)
  - The filter bar's `X` button is now visible again and also clears filters (#1025)
+ - Reverting changes for episodes did not always work (#1032)
 
 ### Changes